# openapi.yaml
openapi: 3.1.0
info:
  title: Moulinator API
  version: 0.1.0
  description: |
    REST contract between the NestJS backend and the NextJS frontend.
    This file is the source of truth for both sides; build agents generate
    typed clients from here and validate controllers against it.
servers:
  - url: http://localhost:3001
    description: Local dev
security:
  - bearerAuth: []
paths:
  /auth/signup:
    post:
      summary: Create an account
      security: []
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SignupRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokens" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Unprocessable" }
  /auth/login:
    post:
      summary: Login
      security: []
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokens" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/refresh:
    post:
      summary: Exchange refresh cookie for new access token
      description: |
        Consumes the httpOnly `mou_rt` cookie set by /auth/login, /auth/signup,
        or a prior /auth/refresh. Requires the CSRF-lite `X-Moulinator-Refresh: 1`
        header to mitigate drive-by refresh via embedded resources. Rotates the
        refresh token on every successful call; the previous refresh jti is
        revoked server-side.
      security: []
      tags: [auth]
      parameters:
        - name: X-Moulinator-Refresh
          in: header
          required: true
          schema: { type: string, enum: ["1"] }
      responses:
        "200":
          description: OK (sets new mou_rt cookie in Set-Cookie)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokens" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /webhooks/jenkins:
    post:
      summary: Jenkins → API webhook
      description: See docs/webhook-contract.md. HMAC-verified, idempotency-keyed.
      security: []
      tags: [webhooks]
      parameters:
        - { name: X-Moulinator-Signature, in: header, required: true, schema: { type: string } }
        - { name: X-Moulinator-Idempotency-Key, in: header, required: true, schema: { type: string, format: uuid } }
        - { name: X-Moulinator-Event, in: header, required: true, schema: { type: string, enum: [build_started, heartbeat, build_completed, build_errored] } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - { $ref: "#/components/schemas/BuildStartedEvent" }
                - { $ref: "#/components/schemas/HeartbeatEvent" }
                - { $ref: "#/components/schemas/BuildCompletedEvent" }
                - { $ref: "#/components/schemas/BuildErroredEvent" }
      responses:
        "200": { description: Processed }
        "202": { description: Queued }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Duplicate idempotency key
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "410":
          description: Run already terminal
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422": { $ref: "#/components/responses/Unprocessable" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Missing/invalid credentials
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: Not found
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Conflict:
      description: Conflict
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unprocessable:
      description: Validation error
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, description: Machine code, e.g. "pat_invalid" }
        message: { type: string }
        details:
          type: object
          additionalProperties: true
    SignupRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 10, writeOnly: true }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, writeOnly: true }
    AuthTokens:
      type: object
      required: [access_token, expires_in]
      description: |
        Refresh token is delivered as an httpOnly cookie (`mou_rt`) on the
        Set-Cookie header — it is never included in the JSON body so that
        XSS cannot exfiltrate it. Clients send `credentials: 'include'` on
        refresh + logout requests and rely on the browser to attach the
        cookie.
      properties:
        access_token: { type: string }
        expires_in: { type: integer, description: Seconds }
...
    BuildStartedEvent:
      type: object
      required: [test_run_id, jenkins_build_url, started_at, runner_image_digest, tests_repo_commit_sha]
      properties:
        test_run_id: { type: string, format: ulid }
        jenkins_build_url: { type: string, format: uri }
        started_at: { type: string, format: date-time }
        runner_image_digest: { type: string }
        tests_repo_commit_sha: { type: string }
    HeartbeatEvent:
      type: object
      required: [test_run_id, heartbeat_at]
      properties:
        test_run_id: { type: string, format: ulid }
        heartbeat_at: { type: string, format: date-time }
        stage: { type: string }
    BuildCompletedEvent:
      type: object
      required: [test_run_id, finished_at, cases]
      properties:
        test_run_id: { type: string, format: ulid }
        finished_at: { type: string, format: date-time }
        cases:
          type: array
          items:
            type: object
            required: [name, status, duration_ms]
            properties:
              name: { type: string }
              status: { type: string, enum: [passed, failed, skipped] }
              duration_ms: { type: integer }
              preview: { type: string }
              artifact_ref: { type: string }
        artifacts:
          type: array
          items:
            type: object
            required: [kind, s3_key, size_bytes]
            properties:
              kind: { type: string, enum: [logs, tarball, junit] }
              s3_key: { type: string }
              size_bytes: { type: integer }
    BuildErroredEvent:
      type: object
      required: [test_run_id, finished_at, error]
      properties:
        test_run_id: { type: string, format: ulid }
        finished_at: { type: string, format: date-time }
        error: { type: string }
        detail: { type: string }